+++ src/ui/styled/animations/index.ts
@@ -1,0 +1,1 @@
+export { keyframes } from './keyframes';
